+++ .pre-commit-config.yaml
@@ -4,8 +4,14 @@ repos:
     hooks:
       - id: black
         language_version: python3
+        exclude: |
+          (?x)^(tests/conftest.py |
+          )$
   - repo: https://gitlab.com/pycqa/flake8
     rev: 3.9.2
     hooks:
       - id: flake8
         language_version: python3
+        exclude: |
+          (?x)^(tests/conftest.py |
+          )$
