+++ services/email.py
@@ -1,0 +1,6 @@
+class EmailService:
+    def __init__(self, email):
+        self.email = email
+
+    def send(self):
+        return {"message": f"Email sent to {self.email}"}
